@@ -2,17 +2,21 @@
   "name": "check-links",
   "version": "1.1.8",
   "description": "Robustly checks an array of URLs for liveness.",
-  "main": "index.js",
   "repository": "transitive-bullshit/check-links",
-  "author": "Travis Fischer <travis@automagical.ai>",
+  "author": "Travis Fischer <travis@transitivebullsh.it>",
   "license": "MIT",
-  "reveal": true,
+  "type": "module",
+  "module": "./lib/index.js",
+  "exports": {
+    "default": "./lib/index.js"
+  },
+  "sideEffects": false,
+  "engines": {
+    "node": ">=14.17"
+  },
   "scripts": {
     "docs": "update-markdown-jsdoc --no-markdown-toc --shallow",
     "test": "ava -v && standard"
-  },
-  "engines": {
-    "node": ">=8"
   },
   "keywords": [
     "url",
@@ -27,17 +31,18 @@
     "link-check",
     "url-check"
   ],
-  "devDependencies": {
-    "ava": "^1.2.1",
-    "nock": "^10.0.6",
-    "sinon": "^7.4.2",
-    "standard": "^12.0.1",
-    "update-markdown-jsdoc": "^1.0.6"
-  },
   "dependencies": {
-    "got": "^11.8.5",
-    "is-relative-url": "^2.0.0",
-    "p-map": "^2.0.0",
-    "p-memoize": "^2.1.0"
+    "got": "^12.5.2",
+    "is-relative-url": "^4.0.0",
+    "p-map": "^5.5.0",
+    "p-memoize": "^7.1.1"
+  },
+  "devDependencies": {
+    "ava": "^5.0.1",
+    "nock": "^13.2.9",
+    "prettier": "^2.7.1",
+    "sinon": "^14.0.2",
+    "standard": "^17.0.0",
+    "update-markdown-jsdoc": "^1.0.6"
   }
 }
