@@ -1,13 +1,12 @@
-'use strict'
+import got from 'got'
+import http from 'http'
+import https from 'https'
+import { utils } from './utils.js'
 
-const got = require('got')
-const isRelativeUrl = require('is-relative-url')
-const parse = require('url').parse
-
-const protocolWhitelist = new Set([
-  'https:',
-  'http:'
-])
+export const agent = {
+  http: new http.Agent({ rejectUnauthorized: false }),
+  https: new https.Agent({ rejectUnauthorized: false })
+}
 
 /**
  * Checks if a URL is alive (2XX status code).
@@ -20,29 +19,29 @@ const protocolWhitelist = new Set([
  *
  * @return {Promise<LivenessResult>}
  */
-module.exports = (url, opts = { }) => {
-  const {
-    headers = { },
-    ...rest
-  } = opts
+export function checkLink(url, opts = {}) {
+  const { headers = {}, ...rest } = opts
 
   opts = {
     headers: {
-      'user-agent': module.exports.userAgent,
+      'user-agent': utils.userAgent,
       'Upgrade-Insecure-Requests': 1,
-      'connection': 'keep-alive',
-      'accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
+      connection: 'keep-alive',
+      accept:
+        'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
       'accept-encoding': 'gzip, deflate, br',
       'cache-control': 'max-age=0',
       'accept-language': 'en-US,en;q=0.9',
       ...headers
     },
-    rejectUnauthorized: false,
-    timeout: 30000,
+    agent,
+    timeout: {
+      request: 30000
+    },
     ...rest
   }
 
-  if (!module.exports.isValidUrl(url, opts)) {
+  if (!utils.isValidUrl(url, opts)) {
     return Promise.resolve({
       status: 'invalid'
     })
@@ -50,43 +49,37 @@ module.exports = (url, opts = { }) => {
 
   // We only allow retrying the GET request because we don't want to wait on
   // exponential falloff for failed HEAD request and failed GET requests.
-  const fetchHEAD = () => (
-    got.head(url, Object.assign({ }, opts, {
-      retry: { retries: 0 }
-    }))
+  const fetchHEAD = () =>
+    got
+      .head(
+        url,
+        Object.assign({}, opts, {
+          retry: { limit: 0 }
+        })
+      )
       .then((res) => ({
         status: 'alive',
         statusCode: res.statusCode
       }))
-  )
 
-  const fetchGET = () => (
-    got(url, opts)
-      .then((res) => ({
-        status: 'alive',
-        statusCode: res.statusCode
-      }))
-  )
+  const fetchGET = () =>
+    got(url, opts).then((res) => ({
+      status: 'alive',
+      statusCode: res.statusCode
+    }))
 
   return fetchHEAD()
-    .catch(() => {
+    .catch((err) => {
+      console.warn('HEAD error', err)
       // TODO: if HEAD results in a `got.HTTPError`, are there status codes where
       // we can bypass the GET request?
       return fetchGET()
     })
-    .catch((err) => ({
-      status: 'dead',
-      statusCode: err.statusCode
-    }))
+    .catch((err) => {
+      console.warn('GET error', err)
+      return {
+        status: 'dead',
+        statusCode: err.statusCode
+      }
+    })
 }
-
-module.exports.isValidUrl = (url, opts) => {
-  if (isRelativeUrl(url)) {
-    return !!opts.baseUrl
-  } else {
-    const parsedUrl = parse(url)
-    return protocolWhitelist.has(parsedUrl.protocol)
-  }
-}
-
-module.exports.userAgent = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/67.0.3396.99 Safari/537.36'
