@@ -1,9 +1,11 @@
-'use strict'
+import pMap from 'p-map'
+import pMemoize from 'p-memoize'
 
-const pMap = require('p-map')
-const pMemoize = require('p-memoize')
+import { checkLink } from './check-link.js'
 
-const checkLink = require('./lib/check-link')
+const isUrlAlive = pMemoize(checkLink, {
+  maxAge: 60 * 1000
+})
 
 /**
  * Robustly checks an array of URLs for liveness.
@@ -31,25 +33,20 @@ const checkLink = require('./lib/check-link')
  *
  * @return {Promise<LivenessResult>}
  */
-module.exports = (urls, opts) => {
-  opts = opts || { }
-
+export async function checkLinks(urls, opts = {}) {
   const concurrency = opts.concurrency || 8
-  const results = { }
+  const results = {}
 
-  return (
-    pMap(urls, (url) => (
-      isUrlAlive(url, opts)
-        .then((result) => {
-          results[url] = result
-        })
-    ), {
+  await pMap(
+    urls,
+    async (url) => {
+      const result = await isUrlAlive(url, opts)
+      results[url] = result
+    },
+    {
       concurrency
-    })
-      .then(() => results)
+    }
   )
-}
 
-const isUrlAlive = pMemoize(checkLink, {
-  maxAge: 60 * 1000
-})
+  return results
+}
