@@ -1,29 +1,28 @@
-'use strict'
+import test from 'ava'
+import sinon from 'sinon'
 
-const test = require('ava')
-const sinon = require('sinon')
-
-const checkLink = require('../lib/check-link')
+import { checkLink, utils } from '../lib/index.js'
 
 test.serial('check-links default options', async (t) => {
-  const stub = sinon.stub(checkLink, 'isValidUrl')
-    .callsFake((url, opts) => {
-      t.is(url, 'invalid')
+  const stub = sinon.stub(utils, 'isValidUrl').callsFake((url, opts) => {
+    const { agent, ...rest } = opts
 
-      t.deepEqual(opts, {
-        headers: {
-          'user-agent': checkLink.userAgent,
-          'Upgrade-Insecure-Requests': 1,
-          'connection': 'keep-alive',
-          'accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
-          'accept-encoding': 'gzip, deflate, br',
-          'cache-control': 'max-age=0',
-          'accept-language': 'en-US,en;q=0.9'
-        },
-        rejectUnauthorized: false,
-        timeout: 30000
-      })
+    t.is(url, 'invalid')
+
+    t.deepEqual(rest, {
+      headers: {
+        'user-agent': utils.userAgent,
+        'Upgrade-Insecure-Requests': 1,
+        connection: 'keep-alive',
+        accept:
+          'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
+        'accept-encoding': 'gzip, deflate, br',
+        'cache-control': 'max-age=0',
+        'accept-language': 'en-US,en;q=0.9'
+      },
+      timeout: { request: 30000 }
     })
+  })
 
   await checkLink('invalid')
 
@@ -31,30 +30,32 @@ test.serial('check-links default options', async (t) => {
 })
 
 test.serial('check-links overriding got options', async (t) => {
-  const stub = sinon.stub(checkLink, 'isValidUrl')
-    .callsFake((url, opts) => {
-      t.deepEqual(opts, {
-        headers: {
-          'user-agent': checkLink.userAgent,
-          'Upgrade-Insecure-Requests': 1,
-          'connection': 'keep-alive',
-          'accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
-          'accept-encoding': 'gzip, deflate, br',
-          'cache-control': 'max-age=0',
-          'accept-language': 'en-US,en;q=0.9',
-          'authorization': 'test'
-        },
-        rejectUnauthorized: false,
-        timeout: 30000,
-        retry: 3
-      })
+  const stub = sinon.stub(utils, 'isValidUrl').callsFake((url, opts) => {
+    const { agent, ...rest } = opts
+
+    t.deepEqual(rest, {
+      headers: {
+        'user-agent': utils.userAgent,
+        'Upgrade-Insecure-Requests': 1,
+        connection: 'keep-alive',
+        accept:
+          'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
+        'accept-encoding': 'gzip, deflate, br',
+        'cache-control': 'max-age=0',
+        'accept-language': 'en-US,en;q=0.9',
+        authorization: 'test'
+      },
+      timeout: { request: 10000 },
+      retry: { limit: 5 }
     })
+  })
 
   await checkLink('invalid', {
     headers: {
-      'authorization': 'test'
+      authorization: 'test'
     },
-    retry: 3
+    timeout: { request: 10000 },
+    retry: { limit: 5 }
   })
 
   stub.restore()
